test(gui): fix and enable multi-account test scenario (#869)
* test: try to delete config dir only if exists


* test: fix config path


* test: check for nul values


* test: enable multiple account test


* test: fix multiple account test


* test: add sure assertion module


* test: use const for config filename


* test: fix python requirement entry


---------

+++ test/gui/environment.py
@@ -18,24 +18,24 @@ def before_scenario(context, feature):
 
 
 def after_scenario(context, scenario):
-    # clean up config dir
-    shutil.rmtree(get_config("clientConfigDir"))
     # clean up sync dir
-    for entry in os.scandir(get_config("clientRootSyncPath")):
-        try:
-            if entry.is_file() or entry.is_symlink():
-                print("Deleting file: " + entry.name)
-                os.unlink(prefix_path_namespace(entry.path))
-            elif entry.is_dir():
-                print("Deleting folder: " + entry.name)
-                shutil.rmtree(prefix_path_namespace(entry.path))
-        except OSError as e:
-            print(f"Failed to delete '{entry.name}'.\nReason: {e}.")
+    if os.path.exists(get_config("clientRootSyncPath")):
+        for entry in os.scandir(get_config("clientRootSyncPath")):
+            try:
+                if entry.is_file() or entry.is_symlink():
+                    print("Deleting file: " + entry.name)
+                    os.unlink(prefix_path_namespace(entry.path))
+                elif entry.is_dir():
+                    print("Deleting folder: " + entry.name)
+                    shutil.rmtree(prefix_path_namespace(entry.path))
+            except OSError as e:
+                print(f"Failed to delete '{entry.name}'.\nReason: {e}.")
     # cleanup paths created outside of the temporary directory during the test
     cleanup_created_paths()
     delete_created_users()
     # quit the application
-    app().quit()
+    if app() is not None:
+        app().quit()
     for process in psutil.process_iter(['pid', 'exe']):
         if process.info['exe'] == get_config("app_path"):
             print("Closing desktop client...")
